git-svn-id: https://svn.resiprocate.org/rep/resiprocate/branches/b-gsoc-2013@10534 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ resip/stack/TransactionState.hxx
@@ -179,7 +179,7 @@ class TransactionState : public DnsHandler
       TransportFailure::FailureReason mFailureReason;      
       int mFailureSubCode;
 
-      static unsigned long StatelessIdCounter;
+      static UInt32 StatelessIdCounter;
       
       friend EncodeStream& operator<<(EncodeStream& strm, const TransactionState& state);
       friend class TransactionController;
